# swagger.yaml
openapi: 3.0.3
info:
  title: InfoHub API - Sistema Completo
  description: |
    API completa do InfoHub para gerenciamento de usuários, produtos, categorias e sistema de recuperação de senhas.
    
    ## Arquitetura
    - **API Principal**: Node.js/Express (porta 8080)
    - **Banco**: MySQL com Prisma ORM
    
    ## Autenticação
    Rotas protegidas requerem JWT token no header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Serviços
    1. **Usuários**: Cadastro, login, CRUD
    2. **Produtos**: Gerenciamento completo de produtos
    3. **Categorias**: Gerenciamento de categorias de produtos
    4. **Recuperação de Senha**: Solicitação, validação, redefinição
    
  version: 1.0.0
  contact:
    name: InfoHub Team
    email: support@infohub.com
  license:
    name: ISC
servers:
  - url: http://localhost:8080/v1/infohub
    description: API Principal (Node.js)
...
components:
  schemas:
...
    EstabelecimentoRequest:
      type: object
      required:
        - nome
        - cnpj
      properties:
        nome:
          type: string
          description: Nome do estabelecimento
          example: "Supermercado Central"
        cnpj:
          type: string
          description: CNPJ do estabelecimento
          example: "12.345.678/0001-90"
        telefone:
          type: string
          description: Telefone do estabelecimento
          example: "(11) 3333-4444"
    EstabelecimentoResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        status_code:
          type: integer
          example: 201
        message:
          type: string
          example: "Estabelecimento criado com sucesso"
        id:
          type: integer
          example: 1
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /estabelecimento:
    post:
      tags:
        - Estabelecimentos
      summary: Criar novo estabelecimento
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstabelecimentoRequest'
      responses:
        '201':
          description: Estabelecimento criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstabelecimentoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /estabelecimento/{id}:
    put:
      tags:
        - Estabelecimentos
      summary: Atualizar estabelecimento por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstabelecimentoRequest'
      responses:
        '200':
          description: Estabelecimento atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstabelecimentoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Estabelecimentos
      summary: Deletar estabelecimento por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Estabelecimento deletado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstabelecimentoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Estabelecimentos
      summary: Buscar estabelecimento por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Estabelecimento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstabelecimentoDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /estabelecimentos:
    get:
      tags:
        - Estabelecimentos
      summary: Listar todos os estabelecimentos
      responses:
        '200':
          description: Lista de estabelecimentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstabelecimentosListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /endereco-usuario:
    post:
      tags:
        - Endereços de Usuário
      summary: Criar novo endereço de usuário
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnderecoUsuarioRequest'
      responses:
        '201':
          description: Endereço criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoUsuarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /endereco-usuario/{id}:
    put:
      tags:
        - Endereços de Usuário
      summary: Atualizar endereço de usuário por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnderecoUsuarioRequest'
      responses:
        '200':
          description: Endereço atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoUsuarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Endereços de Usuário
      summary: Deletar endereço de usuário por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Endereço deletado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoUsuarioResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Endereços de Usuário
      summary: Buscar endereço de usuário por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Endereço encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecoUsuarioDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /enderecos-usuario:
    get:
      tags:
        - Endereços de Usuário
      summary: Listar todos os endereços de usuário
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de endereços de usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecosUsuarioListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /usuario/{id}/enderecos:
    get:
      tags:
        - Endereços de Usuário
      summary: Buscar endereços por ID do usuário
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Endereços do usuário encontrados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnderecosUsuarioListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /categoria:
    post:
      tags:
        - Categorias
      summary: Criar nova categoria
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '201':
          description: Categoria criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /categoria/{id}:
    put:
      tags:
        - Categorias
      summary: Atualizar categoria por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoriaRequest'
      responses:
        '200':
          description: Categoria atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Categorias
      summary: Deletar categoria por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Categoria deletada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Categorias
      summary: Buscar categoria por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Categoria encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /categorias:
    get:
      tags:
        - Categorias
      summary: Listar todas as categorias
      responses:
        '200':
          description: Lista de categorias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriasListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /categorias-produtos:
    get:
      tags:
        - Categorias
      summary: Listar todas as categorias com seus produtos
      responses:
        '200':
          description: Lista de categorias com produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriasComProdutosResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /categoria-produtos/{id}:
    get:
      tags:
        - Categorias
      summary: Buscar categoria por ID com seus produtos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Categoria encontrada com produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoriaComProdutosDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /produtos:
    post:
      tags:
        - Produtos
      summary: Criar novo produto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoRequest'
      responses:
        '201':
          description: Produto criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Produtos
      summary: Listar todos os produtos
      responses:
        '200':
          description: Lista de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutosListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /produto/{id}:
    put:
      tags:
        - Produtos
      summary: Atualizar produto por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProdutoRequest'
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Produtos
      summary: Deletar produto por ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto deletado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Produtos
      summary: Buscar produto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProdutoDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==============================
  # API PRINCIPAL (Node.js - Porta 8080)
  # ==============================
  
  # USUÁRIOS
  /usuarios/cadastro:
    post:
      tags:
        - Usuários
      summary: Cadastrar novo usuário
      description: Cria um novo usuário no sistema
      servers:
        - url: http://localhost:8080/v1/infohub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
            example:
              nome: "João Silva"
              email: "joao@email.com"
              senha: "123456"
              data_nascimento: "1990-05-15"
              telefone: "(11) 99999-9999"
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email já existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /usuarios:
    get:
      tags:
        - Usuários
      summary: Listar todos os usuários
      description: Retorna lista de todos os usuários cadastrados
      servers:
        - url: http://localhost:8080/v1/infohub
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Usuario'
        '401':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /login:
    post:
      tags:
        - Autenticação
      summary: Fazer login
      description: Autentica usuário e retorna JWT token
      servers:
        - url: http://localhost:8080/v1/infohub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # RECUPERAÇÃO DE SENHA
...
  /respostaOllama:
    post:
      tags:
        - Chat IA
      summary: Resposta Ollama (Legacy)
      description: Endpoint legacy para compatibilidade. Use /interagir para novas implementações.
      servers:
        - url: http://localhost:8080/v1/infohub
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Resposta processada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'